user-story-tests/15-inherit-resets-layer-as-lower-priority-renamed-layer.html

<!DOCTYPE html>
<html lang="en">

<head>
    <meta charset="UTF-8">
    <meta name="viewport" content="width=device-width, initial-scale=1.0">
    <title>Inherit resets layer as lower priority renamed layer</title>
    <script src='./resources/testharness.js'></script>
    <script src='./resources/testharnessreport.js'></script>
</head>

<body>

    <style>
        @layer resets {
            button {
                border: thick dashed red;
            }
        }
    </style>
    <button>Button outside a shadow tree</button>
    <button-group>
        <template shadowrootmode="open">
            <style>
                @layer inherit.resets.as.shadowresets, shadowresets, resets;

                @layer resets {
                    button {
                        border: thick solid black;
                    }
                }
            </style>
            <button>Button inside a shadow tree (thick solid black) (from shadow tree resets)</button>
        </template>
    </button-group>


    <script type="module">
        import ShadowLayers from '../ShadowLayers.js';

        window.onload = function () {

            const elName = "button-group"

            test(function () {

                const inheritResult = ShadowLayers.inheritFromShadowStatementRule(document.querySelector(elName))
                assert_true(inheritResult, "inherit found, so returns true");
            }, "Inherit resets layer as lower priority renamed layer");

        }
    </script>

    <div id="log"></div>

</body>

</html>
